+++ test/flake-parts-override-default-cmd/flake.nix
@@ -40,8 +40,6 @@
         in
         {
           conan = {
-            settings.base = { };
-
             defaults.enable = true;
 
             inherit configLocal conanHome buildType compilerCppStd compilerLibCxx;